Changes Made to Servicing
Previously, whether or not a vehicle needed to be serviced was determined by dividing its odometer reading by the regular service interval (10,000), and checking that the recorded service count equaled this figure. It is now determined by taking the difference between the current odometer reading and the odometer reading from the last service, and seeing if this figure is greater than the regular service interval. This change was made because if a car had traveled more than double the distance of the recommended service interval and then been serviced, it would still need to be serviced immediately. Doing it this way makes more sense.